refactor: 클린 아키텍처 폴더 구조에 맞춰 인증 시스템 재구성
- 도메인 계층에서 애플리케이션 계층으로 에러 클래스 이동
- 유스케이스별로 개별 클래스 분리 (RegisterUseCase, LoginUseCase 등)
- 컨벤션에 따라 application/dto 폴더에 에러 DTO 클래스 배치
- 모든 테스트가 새로운 구조에서 정상 동작하는지 확인 완료
