+++ app/api/auth/application/dto/AuthError.ts
@@ -1,0 +1,7 @@
+// 인증 도메인 특화 에러들의 기본 클래스
+export class AuthError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'AuthError';
+  }
+}
