+++ app/api/auth/application/dto/InvalidNameError.ts
@@ -1,0 +1,8 @@
+import { AuthError } from './AuthError';
+
+export class InvalidNameError extends AuthError {
+  constructor(message: string = '이름은 필수입니다.') {
+    super(message);
+    this.name = 'InvalidNameError';
+  }
+}
