+++ app/api/auth/application/dto/index.ts
@@ -1,0 +1,14 @@
+// Request/Response DTOs
+export { SignupRequest } from './SignupRequest';
+export { SignupResponse } from './SignupResponse';
+export { LoginRequest } from './LoginRequest';
+export { EmailCheckResponse } from './EmailCheckResponse';
+
+// Error DTOs
+export { AuthError } from './AuthError';
+export { DuplicateEmailError } from './DuplicateEmailError';
+export { InvalidCredentialsError } from './InvalidCredentialsError';
+export { WeakPasswordError } from './WeakPasswordError';
+export { InvalidEmailFormatError } from './InvalidEmailFormatError';
+export { InvalidNameError } from './InvalidNameError';
+export { ValidationError } from './ValidationError';
