+++ app/api/auth/application/usecases/index.ts
@@ -1,0 +1,4 @@
+export { RegisterUseCase } from './RegisterUseCase';
+export { LoginUseCase } from './LoginUseCase';
+export { VerifyCredentialsUseCase } from './VerifyCredentialsUseCase';
+export { CheckEmailDuplicationUseCase } from './CheckEmailDuplicationUseCase';
